+++ CHANGELOG.md
@@ -16,17 +16,12 @@ Types of changes:
 
 ### ➕  New Features 
 
-- Added Adaptive Profile Support as defined by the QIR Specification for QASM conversions ([#225](https://github.com/qBraid/qbraid-qir/pull/225))
-- Edited Custom Profile to Base Profile as defined by the QIR Specification for QASM conversions ([#228](https://github.com/qBraid/qbraid-qir/pull/228))
-
 ### 🌟  Improvements
-- Improved coverage for `qbraid_qir` with additional tests for the gate maps for QASM 3 conversions. ( [#227](https://github.com/qBraid/qbraid-qir/pull/227) )
 
 ### 📜  Documentation
 
 ### 🐛  Bug Fixes
 
 ### ⬇️  Dependency Updates 
-- Bump `pyqasm` version from `>=0.3.2, <0.4.0` to `>=0.4.0, <0.5.0`. ([#231](https://github.com/qBraid/qbraid-qir/pull/231))
 
 ### 👋  Deprecations
