+++ CITATION.cff
@@ -6,20 +6,11 @@ authors:
 - given-names: Harshit
   family-names: Gupta
   affiliation: qBraid Co.
-- given-names: Rohan
-  family-names: Jain
-  affiliation: Carnegie Mellon University
-- given-names: Samuel
-  family-names: Kushnir
-  affiliation: University of Maryland
-- given-names: Priyansh
-  family-names: Parakh
-  affiliation: University of Toronto
 - given-names: Ryan James
   family-names: Hill
   affiliation: qBraid Co.
 repository-code: https://github.com/qBraid/qbraid-qir
-url: https://qbraid.github.io/qbraid-qir
+url: https://sdk.qbraid.com/qbraid-qir
 repository-artifact: https://github.com/qBraid/qbraid-qir/releases/tag/v0.4.0
 keywords:
 - python
