+++ docs/requirements.txt
@@ -1,0 +1,5 @@
+sphinx>=7.3.7,<=8.3.0
+sphinx-autodoc-typehints>=1.24,<3.2
+sphinx-rtd-theme>=2.0,<3.1
+docutils<0.22
+sphinx-copybutton
