+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61", "wheel"]
+requires = ["setuptools>=77", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -8,7 +8,7 @@ dynamic = ["version"]
 description = "qBraid-SDK extension providing support for QIR conversions."
 readme = "README.md"
 authors = [{name = "qBraid Development Team"}, {email = "contact@qbraid.com"}]
-license = {text = "Apache 2.0"}
+license = "Apache-2.0"
 keywords = ["qbraid", "quantum", "qir", "llvm", "cirq", "openqasm"]
 classifiers = [
     "Development Status :: 5 - Production/Stable",
@@ -40,8 +40,6 @@ Discord = "https://discord.gg/TPBU2sa8Et"
 cirq = ["cirq-core>=1.3.0,<1.6.0"]
 qasm3 = ["pyqasm>=0.4.0,<0.5.0", "numpy"]
 test = ["qbraid>=0.9.0,<0.10.0", "pytest", "pytest-cov", "autoqasm>=0.1.0"]
-lint = ["black", "isort", "pylint", "qbraid-cli>=0.10.1"]
-docs = ["sphinx>=7.3.7,<=8.3.0", "sphinx-autodoc-typehints>=1.24,<3.2", "sphinx-rtd-theme>=2.0,<3.1", "docutils<0.22", "sphinx-copybutton"]
 
 [tool.setuptools]
 packages = ["qbraid_qir", "qbraid_qir.cirq", "qbraid_qir.qasm3"]
@@ -55,7 +53,7 @@ disable = "C0103, C0115, C0116, I1101, R0903, R0917, W0212, W0511"
 
 [tool.black]
 line-length = 100
-target-version = ["py310", "py311", "py312"]
+target-version = ["py310", "py311", "py312", "py313"]
 
 [tool.isort]
 profile = "black"
