+++ tox.ini
@@ -22,8 +22,7 @@ commands =
 
 [testenv:docs]
 description = Use sphinx to build the HTML docs.
-extras =
-    docs
+deps = -r {toxinidir}/docs/requirements.txt
 commands =
     sphinx-build -W -b html docs/ docs/build/html {posargs}
